frontend-next/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.24.0
sqlalchemy==2.0.24
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.1.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
cryptography==41.0.5
loguru==0.7.2
pytest-asyncio==0.23.2
fpdf2==2.7.7
pandas==2.2.0
openpyxl==3.1.2
asyncpg==0.29.0
requests==2.31.0
PyJWT==2.8.0
email-validator==2.2.0